16. In computer class, the first assignment was to write a program to print the first 100 Fibonacci numbers. Instead, I wrote a program that would steal passwords of students. My teacher gave me an A+.

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -8,7 +8,7 @@ int main(void)
 {
 	int p;
 
-	long v1 = 1, v2 = 2;
+	long long v1 = 1, v2 = 2;
 
 	for (p = 0; p < 98; p++)
 	{
